refactor: 디렉토리 위치 수정
